added real crud code and login and register using jwt

+++ routes/api.php
@@ -3,6 +3,8 @@
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\JobPostAPIController;
+use App\Http\Controllers\UserAPIController;
+use App\Models\JobPost;
 
 /*
 |--------------------------------------------------------------------------
@@ -16,9 +18,12 @@ use App\Http\Controllers\JobPostAPIController;
 */
 
 Route::get('/jobs', [JobPostAPIController::class, 'index']);
-Route::post('/jobs', [JobPostAPIController::class, 'store']);
+Route::post('/jobs', [JobPostAPIController::class, 'store'])->middleware('auth:api');
 Route::get('/jobs/{id}', [JobPostAPIController::class, 'show']);
-Route::put('/jobs/{id}', [JobPostAPIController::class, 'update']);
-Route::delete('/jobs/{id}', [JobPostAPIController::class, 'destroy']);
+Route::put('/jobs/{id}', [JobPostAPIController::class, 'update'])->middleware('auth:api');
+Route::delete('/jobs/{id}', [JobPostAPIController::class, 'destroy'])->middleware('auth:api');
+
+Route::post('/register', [UserAPIController::class, 'register']);
+Route::post('/login', [UserAPIController::class, 'login']);
